+++ bindingcollectionadapter/src/main/java/me/tatarka/bindingcollectionadapter/BindingViewPagerAdapter.java
@@ -50,6 +50,14 @@ public class BindingViewPagerAdapter<T> extends PagerAdapter implements BindingC
         this.selector = selector;
     }
 
+    /**
+     * Constructs a new instance with the given {@link ItemViewArg}.
+     */
+    public BindingViewPagerAdapter(@NonNull ItemViewArg<T> arg) {
+        this.itemView = arg.itemView;
+        this.selector = arg.selector;
+    }
+
     @Override
     public void setItems(@Nullable Collection<T> items) {
         if (this.items == items) {